Link Nvidia Physx (See description)
Built from source and statically linked.
Tested on Linux 6.9.6.

The GPU object at lib/physx-5.4.0/source/compiler/cmake/linux/CMak
eLists.txt has been disabled on line 147 because it requires lib/p
hysx-5.4.0/generate_projects.sh to be run to create that object
binary. This would limit the portability of the code. I think that
GPU-acceleration won't be relevant to this project's simple physics
but I will either bite the bullet or find a way to dynamically
generate it later if this becomes a problem.

Also see lib/physx-5.4.0/documentation.