Add triangle-based collision
Custom objects can now have collision applied to their mesh.

Problem: The collision mesh is only applied to one side of each
triangle in PxCookTriangleMesh() and the sidedness looks to be
determined by the order of allIndices. The quick solution of this
was to apply eFLIPNORMALS and just cook the mesh twice, meaning
both the inside and outside of meshes will have collision. In
future though I would like to figure out how to read in indices in
the correct order so that the collision mesh is only pointing
outwards and the function is only called once. This will probably
have to be done by using the normals for each vertex.