Bound shadowmap tightly around camera
The calcProjectedLightSpace function has been moved to camera.cpp
and is responsible for moving the depthmap.
The near/far distance determines how near/far the map will bound
around the camera. E.g., if the near/far plane of the camera is
0 and 10, then calculating the PLS with the near/far distances of
0 and 5 would show shadows in the half of world space closest to
the camera, and not the rest. This will be used for cascading shadow
maps next.